Fixed double-authentication issue
90% sure this fixes the issue where double clicking the login button when the window is closing causes an error

+++ core/windows/loadingWindowHandler.js
@@ -62,20 +62,22 @@ class LoadingWindowHandler extends WindowHandler {
                                     // (Basically, whether the server is online and functioning properly or not)
                                     if (!res.serverError) {
 
-                                        this.window.hide();
-                
-                                        setTimeout(() => {
+                                        if (this.window) {
+                                            this.window.hide();
+                    
+                                            setTimeout(() => {
 
-                                            if (res.verified) {
-                                                WindowController.spawnWindow("mainWindowHandler");
-                                            } else {
-                                                db.remove(userInfodb, "token");
-                                                WindowController.spawnWindow("loginWindowHandler");
-                                            }
+                                                if (res.verified) {
+                                                    WindowController.spawnWindow("mainWindowHandler");
+                                                } else {
+                                                    db.remove(userInfodb, "token");
+                                                    WindowController.spawnWindow("loginWindowHandler");
+                                                }
 
-                                            this.window.close();   
-                                            resolve(true); 
-                                        }, 1000);
+                                                this.window.close();   
+                                                resolve(true); 
+                                            }, 1000);
+                                        }
 
                                     } else {
                                         // TODO: Handle server error
